requirements.txt
chromadb==0.4.22
langchain==0.1.0
langchain-community==0.0.20
langchain-google-genai==0.0.6
sentence-transformers==2.7.0
huggingface-hub==0.23.0
streamlit==1.28.0
pypdf2==3.0.1
python-docx==0.8.11
google-generativeai==0.3.2
pandas==2.1.4
openpyxl==3.1.2
python-dotenv==1.0.0
